Move ViaVai to user pages and remove legacy Form/Table
