--- web/src/types/viavai/sidebar.types.ts
@@ -1,12 +1,0 @@
-import { type ReactNode } from 'react';
-
-export type SidebarItem = {
-    name: string;
-    icon: string;
-    element: ReactNode;
-};
-
-export type SidebarSchemaConfig = {
-    title?: string;
-    items: SidebarItem[];
-};
